+++ org-trello-tests.el
@@ -231,21 +231,15 @@
   (expect (format "%s%s" *TRELLO-URL* "/uri/other")      (orgtrello-query/--compute-url *TRELLO-URL* "/uri/other"))
   (expect (format "some-server/uri/some/other")          (orgtrello-query/--compute-url "some-server" "/uri/some/other")))
 
-(defvar org-trello-tests/--query-map (make-hash-table :test 'equal))
-(puthash :method :some-get org-trello-tests/--query-map)
-(puthash :uri    :some-uri org-trello-tests/--query-map)
-(puthash :sync   :some-sync org-trello-tests/--query-map)
-(puthash :params :some-params org-trello-tests/--query-map)
+(expectations
+  (expect :some-get (orgtrello-query/--method (orgtrello-hash/make-properties `((:method . :some-get) (:uri . :some-uri) (:sync . :some-sync) (:params . :some-params)))))
+  (expect :some-uri (orgtrello-query/--uri (orgtrello-hash/make-properties `((:method . :some-get) (:uri . :some-uri) (:sync . :some-sync) (:params . :some-params)))))
+  (expect :some-sync (orgtrello-query/--sync (orgtrello-hash/make-properties `((:method . :some-get) (:uri . :some-uri) (:sync . :some-sync) (:params . :some-params)))))
+  (expect :some-params (orgtrello-query/--params (orgtrello-hash/make-properties `((:method . :some-get) (:uri . :some-uri) (:sync . :some-sync) (:params . :some-params))))))
 
 (expectations
-  (expect :some-get (orgtrello-query/--method org-trello-tests/--query-map))
-  (expect :some-uri (orgtrello-query/--uri org-trello-tests/--query-map))
-  (expect :some-sync (orgtrello-query/--sync org-trello-tests/--query-map))
-  (expect :some-params (orgtrello-query/--params org-trello-tests/--query-map)))
-
-(expectations
-  (expect :some-id (orgtrello-query/--id '((id . :some-id))))
-  (expect nil      (orgtrello-query/--id '((noid . :some-id)))))
+  (expect "some-id" (orgtrello-query/--id '((id . "some-id"))))
+  (expect nil       (orgtrello-query/--id '((noid . "some-id")))))
 
 (expectations
   (expect :some-name (orgtrello-query/--name '((name . :some-name))))
@@ -282,8 +276,8 @@
 ;; ########################## orgtrello-tests
 
 (ert-deftest testing-orgtrello/--compute-data-from-entity-meta ()
-  (let* ((entry   (orgtrello-hash/make-hash-org :some-level :some-keyword :some-name :some-id :some-due :some-point :some-buffername)))
-    (should (equal (orgtrello/--id entry)         :some-id))
+  (let* ((entry   (orgtrello-hash/make-hash-org :some-level :some-keyword :some-name "some-id" :some-due :some-point :some-buffername)))
+    (should (equal (orgtrello/--id entry)         "some-id"))
     (should (equal (orgtrello/--name entry)       :some-name))
     (should (equal (orgtrello/--keyword entry)    :some-keyword))
     (should (equal (orgtrello/--level entry)      :some-level))
@@ -385,7 +379,7 @@
  (expect nil                            (orgtrello-proxy/--dispatch-action "nothing")))
 
 (expectations
-  (expect :id                                                         (orgtrello/--compute-marker-from-entry (orgtrello-hash/make-hash-org :level :kwd :name      :id  :due :position :buffername)))
+  (expect "id"                                                        (orgtrello/--compute-marker-from-entry (orgtrello-hash/make-hash-org :level :kwd :name      "id"  :due :position :buffername)))
   (expect "orgtrello-marker-2a0b98e652ce6349a0659a7a8eeb3783ffe9a11a" (orgtrello/--compute-marker-from-entry (orgtrello-hash/make-hash-org :level :kwd "some-name" nil :due 1234      "buffername")))
   (expect "orgtrello-marker-6c59c5dcf6c83edaeb3f4923bfd929a091504bb3" (orgtrello/--compute-marker-from-entry (orgtrello-hash/make-hash-org :level :kwd "some-name" nil :due 4321      "some-other-buffername"))))
 
@@ -394,8 +388,16 @@
   (expect "orgtrello-marker-6c59c5dcf6c83edaeb3f4923bfd929a091504bb3" (orgtrello/compute-marker "some-other-buffername" "some-name" 4321)))
 
 (expectations
-  (expect "marker" (orgtrello-proxy/--compute-pattern-search-from-marker "marker"))
-  (expect ":orgtrello-marker-tony:" (orgtrello-proxy/--compute-pattern-search-from-marker "orgtrello-marker-tony")))
+  (expect "marker-is-a-trello-id" (orgtrello-proxy/--compute-pattern-search-from-marker "marker-is-a-trello-id"))
+  (expect "orgtrello-marker-tony" (orgtrello-proxy/--compute-pattern-search-from-marker "orgtrello-marker-tony")))
+
+(expectations
+  (expect t   (orgtrello/id-p "anything-that-does-not-start-with-orgtrello-marker"))
+  (expect t   (orgtrello/id-p "agfgdsfgbdfgbdfgbdfshouldbetrue"))
+  (expect t   (orgtrello/id-p "orgtrello-markeragfgdsfgbdfgbdfgbdfshouldbetrue"))
+  (expect t   (orgtrello/id-p "should-be-true-orgtrello-marker-agfgdsfgbdfgbdfgbdf"))
+  (expect nil (orgtrello/id-p "orgtrello-marker-shouldbenil"))
+  (expect nil (orgtrello/id-p nil)))
 
 (expectations
   (expect '(3 5 7) (--map (funcall (compose-fn '((lambda (it) (+ 1 it)) (lambda (it) (* 2 it)))) it) '(1 2 3))))
@@ -468,19 +470,318 @@
      (orgtrello-hash/make-hash-org 1 :kwd :name nil :due :position :buffer-name)
      (lambda (entity s) (format "%S %s" entity s))
      "- hello"))
-  (expect "#s(hash-table size 65 test equal rehash-size 1.5 rehash-threshold 0.8 data (:buffername :buffer-name :position :position :level 1 :keyword :kwd :name :name :id :some-id :due :due)) - hello"
+  (expect "#s(hash-table size 65 test equal rehash-size 1.5 rehash-threshold 0.8 data (:buffername :buffer-name :position :position :level 1 :keyword :kwd :name :name :id \"some-id\" :due :due)) - hello"
 
     (org-action/--functional-controls-then-do
      '(orgtrello/--right-level-p orgtrello/--already-synced-p)
-     (orgtrello-hash/make-hash-org 1 :kwd :name :some-id :due :position :buffer-name)
+     (orgtrello-hash/make-hash-org 1 :kwd :name "some-id" :due :position :buffer-name)
      (lambda (entity s) (format "%S %s" entity s))
      "- hello")))
-
 
 (expectations
   (expect "- message 1\n- message 2\n" (org-action/--compute-error-message '("message 1" "message 2"))))
 
+(expectations
+ (expect ":key:" (orgtrello-hash/key "key")))
+
+(expectations
+  (expect 2 (orgtrello-cbx/--level '("-" "[X]" "call" "people" "[4/4]")))
+  (expect 3 (orgtrello-cbx/--level '("" "-" "[X]" "Peter")))
+  (expect 3 (orgtrello-cbx/--level '("" "" "-" "[X]" "Peter")))
+  (expect 3 (orgtrello-cbx/--level '("" "" "" "-" "[X]" "Peter")))
+  (expect 4 (orgtrello-cbx/--level '("" "" "" "" "" "Peter"))))
+
+(expectations
+  (expect "DONE" (orgtrello-cbx/--status"[X]"))
+  (expect "TODO" (orgtrello-cbx/--status"[ ]"))
+  (expect "TODO" (orgtrello-cbx/--status"[]"))
+  (expect "TODO" (orgtrello-cbx/--status"[-]"))
+  (expect "TODO" (orgtrello-cbx/--status"")))
+
+(expectations
+  (expect '("-" "[X]" "call" "people" "[4/4]")                                             (orgtrello-cbx/--org-split-data "- [X] call people [4/4]"))
+  (expect '("-" "[X]" "call" "people" "[4/4]" ":PROPERTIES:" "{\"orgtrello-id\":\"456\"}") (orgtrello-cbx/--org-split-data "- [X] call people [4/4] :PROPERTIES: {\"orgtrello-id\":\"456\"}"))
+  (expect '("" "" "-" "[X]" "Peter")                                                       (orgtrello-cbx/--org-split-data "  - [X] Peter"))
+  (expect '("" "" "-" "[X]" "Peter" ":PROPERTIES:" "{\"orgtrello-id\":\"456\"}")           (orgtrello-cbx/--org-split-data "  - [X] Peter :PROPERTIES: {\"orgtrello-id\":\"456\"}"))
+
+  (expect '("-" "[]" "call" "people" "[4/4]")                                             (orgtrello-cbx/--org-split-data "- [] call people [4/4]"))
+  (expect '("-" "[]" "call" "people" "[4/4]" ":PROPERTIES:" "{\"orgtrello-id\":\"456\"}") (orgtrello-cbx/--org-split-data "- [] call people [4/4] :PROPERTIES: {\"orgtrello-id\":\"456\"}"))
+  (expect '("" "" "-" "[]" "Peter")                                                       (orgtrello-cbx/--org-split-data "  - [] Peter"))
+  (expect '("" "" "-" "[]" "Peter" ":PROPERTIES:" "{\"orgtrello-id\":\"456\"}")           (orgtrello-cbx/--org-split-data "  - [] Peter :PROPERTIES: {\"orgtrello-id\":\"456\"}"))
+
+  (expect '("-" "[]" "call" "people" "[4/4]")                                             (orgtrello-cbx/--org-split-data "- [ ] call people [4/4]"))
+  (expect '("-" "[]" "call" "people" "[4/4]" ":PROPERTIES:" "{\"orgtrello-id\":\"456\"}") (orgtrello-cbx/--org-split-data "- [ ] call people [4/4] :PROPERTIES: {\"orgtrello-id\":\"456\"}"))
+  (expect '("" "" "-" "[]" "Peter")                                                       (orgtrello-cbx/--org-split-data "  - [ ] Peter"))
+  (expect '("" "" "-" "[]" "Peter" ":PROPERTIES:" "{\"orgtrello-id\":\"456\"}")           (orgtrello-cbx/--org-split-data "  - [ ] Peter :PROPERTIES: {\"orgtrello-id\":\"456\"}")))
+
+(expectations
+  (expect "[X]" (orgtrello-cbx/--retrieve-status '("" "" "-" "[X]" "Peter")))
+  (expect "[]" (orgtrello-cbx/--retrieve-status '("" "" "-" "[]" "Peter")))
+  (expect "[-]" (orgtrello-cbx/--retrieve-status '("" "" "-" "[-]" "Peter")))
+  (expect "[ ]" (orgtrello-cbx/--retrieve-status '("" "" "-" "[ ]" "Peter"))))
+
+(expectations
+  (expect "call people [4/4]" (orgtrello-cbx/--name "- [X] call people [4/4]"   "[X]"))
+  (expect "call people [4/4]" (orgtrello-cbx/--name "- [] call people [4/4]"    "[]"))
+  (expect "call people [4/4]" (orgtrello-cbx/--name "- [-] call people [4/4]"   "[-]"))
+  (expect "call people [4/4]" (orgtrello-cbx/--name "-[X] call people [4/4]"    "[X]"))
+  (expect "call people [4/4]" (orgtrello-cbx/--name "-[] call people [4/4]"     "[]"))
+  (expect "call people [4/4]" (orgtrello-cbx/--name "-[-] call people [4/4]"    "[-]"))
+  (expect "call people [4/4]" (orgtrello-cbx/--name "  - [X] call people [4/4]" "[X]"))
+  (expect "call people [4/4]" (orgtrello-cbx/--name "  - [] call people [4/4]"  "[]"))
+  (expect "call people [4/4]" (orgtrello-cbx/--name "  - [-] call people [4/4]" "[-]"))
+  (expect "call people [4/4]" (orgtrello-cbx/--name "  -[X] call people [4/4]"  "[X]"))
+  (expect "call people [4/4]" (orgtrello-cbx/--name "  -[] call people [4/4]"   "[]"))
+  (expect "call people [4/4]" (orgtrello-cbx/--name "  -[-] call people [4/4]"  "[-]")))
+
+(expectations
+  (expect "{\"orgtrello-id\":\"123\"}"                               (orgtrello-cbx/--to-properties `((,*ORGTRELLO-ID* . "123"))))
+  (expect "{\"orgtrello-id\":\"456\"}"                               (orgtrello-cbx/--to-properties `((,*ORGTRELLO-ID* . "123") (,*ORGTRELLO-ID* . "456"))))
+  (expect "{\"orgtrello-id\":\"def\", \"orgtrello-marker\":\"456\", \"orgtrello-id\":\"abc\"}" (orgtrello-cbx/--to-properties `(("orgtrello-id" . "abc") (orgtrello-marker . "456") (orgtrello-id . "def"))))
+  (expect "{\"orgtrello-marker\":\"456\", \"orgtrello-id\":\"def\"}" (orgtrello-cbx/--to-properties `(("orgtrello-id" . "abc") (orgtrello-marker . "456") ("orgtrello-id" . "def"))))
+  (expect "{\"orgtrello-marker\":\"456\", \"orgtrello-id\":\"def\"}" (orgtrello-cbx/--to-properties `((orgtrello-id . "abc") (orgtrello-marker . "456") (orgtrello-id . "def")))))
+
+(expectations
+  (expect '((orgtrello-id . "123")) (orgtrello-cbx/--from-properties "{\"orgtrello-id\":\"123\"}")))
+
+(expectations
+  (expect '((orgtrello-id . "123")) (orgtrello-cbx/--read-properties "- [X] some checkbox :PROPERTIES: {\"orgtrello-id\":\"123\"}")))
+
+(expectations
+  (expect "- [X] some checkbox :PROPERTIES: {\"orgtrello-id\":\"123\"}"
+    (orgtrello-cbx/--update-properties "- [X] some checkbox :PROPERTIES: {\"orgtrello-id\":\"abc\"}" `((,*ORGTRELLO-ID* . "123"))))
+  (expect "- [X] some checkbox :PROPERTIES: {\"orgtrello-id\":\"456\"}"
+    (orgtrello-cbx/--update-properties "- [X] some checkbox" `((,*ORGTRELLO-ID* . "456"))))
+  (expect "- [X] some checkbox :PROPERTIES: {}"
+    (orgtrello-cbx/--update-properties "- [X] some checkbox" nil))
+  (expect "- [X] some checkbox :PROPERTIES: {\"orgtrello-id\":\"789\"}"
+    (orgtrello-cbx/--update-properties "- [X] some checkbox :PROPERTIES:" `((,*ORGTRELLO-ID* . "789"))))
+  (expect "- [X] some checkbox :PROPERTIES: {}"
+    (orgtrello-cbx/--update-properties "- [X] some checkbox :PROPERTIES:" nil)))
+
+(expectations
+  (expect "123"    (orgtrello-cbx/--org-get-property "orgtrello-id" `((orgtrello-id . "123"))))
+  (expect nil      (orgtrello-cbx/--org-get-property "orgtrello-id" `(("orgtrello-id" . "123"))))
+  (expect nil      (orgtrello-cbx/--org-get-property 'orgtrello-id `(("orgtrello-id" . "123"))))
+  (expect "123"    (orgtrello-cbx/--org-get-property 'orgtrello-id `((orgtrello-id . "123"))))
+  (expect "marker" (orgtrello-cbx/--org-get-property "orgtrello-marker" `(("orgtrello-id" . "123") (orgtrello-marker . "marker")))))
+
+(expectations
+  (expect `((orgtrello-id . "10") (orgtrello-marker . "123"))    (orgtrello-cbx/--org-update-property "orgtrello-id" "10" `((orgtrello-marker . "123"))))
+  (expect `((orgtrello-toto . "abc") (orgtrello-marker . "456")) (orgtrello-cbx/--org-update-property "orgtrello-toto" "abc" `((orgtrello-marker . "456"))))
+  (expect `((orgtrello-id . "abc") (orgtrello-marker . "456"))   (orgtrello-cbx/--org-update-property "orgtrello-id" "abc" `((orgtrello-marker . "456") (orgtrello-id . "def")))))
+
+(expectations
+  (expect `(("orgtrello-id" . "123") (orgtrello-marker . "marker")) (orgtrello-cbx/--org-delete-property "orgtrello-id" `(("orgtrello-id" . "123") (orgtrello-marker . "marker"))))
+  (expect `((orgtrello-marker . "marker")) (orgtrello-cbx/--org-delete-property "orgtrello-id" `((orgtrello-id . "123") (orgtrello-marker . "marker"))))
+  (expect `((orgtrello-marker . "marker")) (orgtrello-cbx/--org-delete-property 'orgtrello-id `((orgtrello-id . "123") (orgtrello-marker . "marker"))))
+  (expect `(("orgtrello-id" . "123") (orgtrello-marker . "marker")) (orgtrello-cbx/--org-delete-property 'orgtrello-id `(("orgtrello-id" . "123") (orgtrello-marker . "marker")))))
+
+(expectations
+  (expect '("  - [X] Peter " " {\"orgtrello-id\":\"456\"}") (orgtrello-cbx/--checkbox-split "  - [X] Peter :PROPERTIES: {\"orgtrello-id\":\"456\"}")))
+
+(expectations
+  (expect "{\"orgtrello-id\":\"456\"}" (orgtrello-cbx/--checkbox-metadata "  - [X] Peter :PROPERTIES: {\"orgtrello-id\":\"456\"}"))
+  (expect ""                           (orgtrello-cbx/--checkbox-metadata "  - [X] Peter :PROPERTIES:"))
+  (expect ""                           (orgtrello-cbx/--checkbox-metadata "  - [X] Peter :PROPERTIES: "))
+  (expect nil                          (orgtrello-cbx/--checkbox-metadata "  - [X] Peter")))
+
+(expectations
+  (expect "  - [X] Peter" (orgtrello-cbx/--checkbox-data "  - [X] Peter :PROPERTIES: {\"orgtrello-id\":\"456\"}")))
+
+(expectations
+  (expect 'some-key (orgtrello-cbx/--key-to-search "some-key"))
+  (expect 'some-key (orgtrello-cbx/--key-to-search 'some-key))
+  (expect :some-key (orgtrello-cbx/--key-to-search :some-key)))
+
+(expectations
+  (expect nil (orgtrello-cbx/--list-is-checkbox-p '("1" "2" "3")))
+  (expect nil (orgtrello-cbx/--list-is-checkbox-p '("1" "2" "3" "-")))
+  (expect t   (orgtrello-cbx/--list-is-checkbox-p '("-" "1" "2" "3")))
+  (expect t   (orgtrello-cbx/--list-is-checkbox-p '("" "" "-" "1" "2" "3"))))
+
+(expectations
+  (expect 1 (orgtrello-cbx/--get-level '(1 2 3)))
+  (expect 2 (orgtrello-cbx/--get-level '(2 3))))
+
+(expectations
+  (expect "DEADLINE: <some-date>
+" (orgtrello/--compute-due-date "some-date"))
+  (expect "" (orgtrello/--compute-due-date nil)))
+
+(expectations
+  (expect "* name TODO
+DEADLINE: <some-date>
+" (orgtrello/--private-compute-card-to-org-entry "TODO" "name" "some-date"))
+  (expect "* name TODO
+" (orgtrello/--private-compute-card-to-org-entry "TODO" "name" nil)))
+
+(expectations
+  (expect "** name\n" (orgtrello/--compute-checklist-to-orgtrello-entry "name"))
+  (expect "** name\n" (orgtrello/--compute-checklist-to-orgtrello-entry "name" nil))
+  (expect "** name\n" (orgtrello/--compute-checklist-to-orgtrello-entry "name" 't))
+  (expect "** name\n" (orgtrello/--compute-checklist-to-orgtrello-entry "name" nil 't))
+  (expect "** name\n" (orgtrello/--compute-checklist-to-orgtrello-entry "name" 't nil))
+  (expect "** name\n" (orgtrello/--compute-checklist-to-orgtrello-entry "name" nil nil))
+  (expect "** name\n" (orgtrello/--compute-checklist-to-orgtrello-entry "name" 't 't)))
+
+(expectations
+  (expect ""      (orgtrello/--symbol " "  0))
+  (expect "*"     (orgtrello/--symbol "*"  1))
+  (expect "****"  (orgtrello/--symbol "**" 2))
+  (expect "   "   (orgtrello/--symbol " "  3)) )
+
+(expectations
+  (expect ""    (orgtrello/--space 0))
+  (expect " "   (orgtrello/--space 1))
+  (expect "  "  (orgtrello/--space 2))
+  (expect "   " (orgtrello/--space 3)) )
+
+(expectations
+  (expect ""    (orgtrello/--star 0))
+  (expect "*"   (orgtrello/--star 1))
+  (expect "**"  (orgtrello/--star 2))
+  (expect "***" (orgtrello/--star 3)) )
+
+(expectations
+  (expect 0 (orgtrello/--compute-level-into-spaces 2))
+  (expect 2 (orgtrello/--compute-level-into-spaces nil))
+  (expect 2 (orgtrello/--compute-level-into-spaces 'any)))
+
+(expectations
+  (expect "- [X] name
+" (orgtrello/--compute-checklist-to-org-checkbox "name" 2 "complete"))
+  (expect "  - [X] name
+" (orgtrello/--compute-checklist-to-org-checkbox "name" 3 "complete"))
+  (expect "- [X] name
+" (orgtrello/--compute-checklist-to-org-checkbox "name" 2 "complete"))
+  (expect "  - [-] name
+" (orgtrello/--compute-checklist-to-org-checkbox "name" 3 "incomplete")))
+
+(expectations
+  (expect "DONE" (orgrello/--compute-item-status "complete"))
+  (expect "TODO" (orgrello/--compute-item-status "anything else will render incomplete")))
+
+(expectations
+  (expect "*** DONE name
+" (orgtrello/--compute-item-to-orgtrello-entry "name" 3 "complete"))
+  (expect "*** TODO name
+" (orgtrello/--compute-item-to-orgtrello-entry "name" 3 "incomplete")))
+
+(expectations
+  (expect "- [-] name
+" (orgtrello/--compute-checklist-to-org-entry `((name . "name")) t))
+  (expect "- [-] name
+" (orgtrello/--compute-checklist-to-org-entry `((name . "name")) t)))
+
+(expectations
+  (expect "** TODO name
+" (orgtrello/--compute-checklist-to-org-entry `((name . "name")) nil))
+  (expect "** TODO name
+" (orgtrello/--compute-checklist-to-org-entry `((name . "name")) nil)))
+
+(expectations
+  (expect "  - [X] name
+" (orgtrello/--compute-item-to-org-entry `((name . "name") (state . "complete")) t))
+  (expect "  - [-] name
+" (orgtrello/--compute-item-to-org-entry `((name . "name") (state . "incomplete")) t)))
+
+(expectations
+  (expect "*** DONE name
+" (orgtrello/--compute-item-to-org-entry `((name . "name") (state . "complete")) nil))
+  (expect "*** TODO name
+" (orgtrello/--compute-item-to-org-entry `((name . "name") (state . "incomplete")) nil)))
+
+(expectations
+  (expect "test" (orgtrello-query/--retrieve-data 'marker `((marker . "test"))))
+  (expect nil (orgtrello-query/--retrieve-data    'other  `((marker . "test"))))
+  (expect "test" (orgtrello-query/--marker                `((marker . "test"))))
+  (expect nil (orgtrello-query/--marker                   `((inexistant . "test"))))
+  (expect "test" (orgtrello-query/--buffername            `((buffername . "test"))))
+  (expect nil (orgtrello-query/--buffername               `((inexistant . "test"))))
+  (expect "test" (orgtrello-query/--position              `((position . "test"))))
+  (expect nil (orgtrello-query/--position                 `((inexistant . "test"))))
+  (expect "test" (orgtrello-query/--id                    `((id . "test"))))
+  (expect nil (orgtrello-query/--id                       `((inexistant . "test"))))
+  (expect "test" (orgtrello-query/--name                  `((name . "test"))))
+  (expect nil (orgtrello-query/--name                     `((inexistant . "test"))))
+  (expect "test" (orgtrello-query/--list-id               `((idList . "test"))))
+  (expect nil (orgtrello-query/--list-id                  `((inexistant . "test"))))
+  (expect "test" (orgtrello-query/--checklist-ids         `((idChecklists . "test"))))
+  (expect nil (orgtrello-query/--checklist-ids            `((inexistant . "test"))))
+  (expect "test" (orgtrello-query/--check-items           `((checkItems . "test"))))
+  (expect nil (orgtrello-query/--check-items              `((inexistant . "test"))))
+  (expect "test" (orgtrello-query/--card-id               `((idCard . "test"))))
+  (expect nil (orgtrello-query/--card-id                  `((inexistant . "test"))))
+  (expect "test" (orgtrello-query/--due                   `((due . "test"))))
+  (expect nil (orgtrello-query/--due                      `((inexistant . "test"))))
+  (expect "test" (orgtrello-query/--state                 `((state . "test"))))
+  (expect nil (orgtrello-query/--state                    `((inexistant . "test"))))
+  (expect "test" (orgtrello-query/--close-property        `((closed . "test"))))
+  (expect nil (orgtrello-query/--close-property           `((inexistant . "test"))))
+  (expect "test" (orgtrello-query/--callback              `((callback . "test"))))
+  (expect nil (orgtrello-query/--callback                 `((inexistant . "test"))))
+  (expect "test" (orgtrello-query/--sync-                 `((sync . "test"))))
+  (expect nil (orgtrello-query/--sync-                    `((inexistant . "test"))))
+  (expect "test" (orgtrello-query/--level                 `((level . "test"))))
+  (expect nil (orgtrello-query/--level                    `((inexistant . "test"))))
+  (expect "test" (orgtrello-query/--method-               `((method . "test"))))
+  (expect nil (orgtrello-query/--method-                  `((inexistant . "test"))))
+  (expect "test" (orgtrello-query/--uri-                  `((uri . "test"))))
+  (expect nil (orgtrello-query/--uri-                     `((inexistant . "test"))))
+  (expect "test" (orgtrello-query/--params-               `((params . "test"))))
+  (expect nil (orgtrello-query/--params-                  `((inexistant . "test"))))
+  (expect "test" (orgtrello-query/--start                 `((start . "test"))))
+  (expect nil (orgtrello-query/--start                    `((inexistant . "test"))))
+  (expect "test" (orgtrello-query/--action                `((action . "test"))))
+  (expect nil (orgtrello-query/--action                   `((inexistant . "test")))))
+
+(expectations
+  (expect :some-method (orgtrello-query/--method (orgtrello-hash/make-properties `((:method . :some-method )))))
+  (expect nil (orgtrello-query/--method (orgtrello-hash/make-properties `((:inexistant . :some-method )))))
+  (expect :some-uri (orgtrello-query/--method (orgtrello-hash/make-properties `((:method . :some-uri )))))
+  (expect nil (orgtrello-query/--method (orgtrello-hash/make-properties `((:inexistant . :some-method )))))
+  (expect :some-sync (orgtrello-query/--method (orgtrello-hash/make-properties `((:method . :some-sync )))))
+  (expect nil (orgtrello-query/--method (orgtrello-hash/make-properties `((:inexistant . :some-method )))))
+  (expect :some-params (orgtrello-query/--method (orgtrello-hash/make-properties `((:method . :some-params )))))
+  (expect nil (orgtrello-query/--method (orgtrello-hash/make-properties `((:inexistant . :some-method ))))))
+
+(expectations
+  (expect "some-method" (orgtrello-query/gethash-data :method (orgtrello-hash/make-properties `((:method . "some-method")))))
+  (expect nil           (orgtrello-query/gethash-data :method (orgtrello-hash/make-properties `((:inexistant . "some-method"))))))
+
+(expectations
+  (expect "DONE" (orgtrello/--compute-state-generic "complete" '("DONE" "TODO")))
+  (expect "TODO" (orgtrello/--compute-state-generic "incomplete" '("DONE" "TODO")))
+
+  (expect "[X]" (orgtrello/--compute-state-generic "complete" '("[X]" "[-]")))
+  (expect "[-]" (orgtrello/--compute-state-generic "incomplete" '("[X]" "[-]"))))
+
+(expectations
+  (expect "[X]" (orgtrello/--compute-state-checkbox "complete"))
+  (expect "[-]" (orgtrello/--compute-state-checkbox "incomplete")))
+
+(expectations
+ (expect "DONE" (orgtrello/--compute-state-item "complete"))
+ (expect "TODO" (orgtrello/--compute-state-item "incomplete")))
 (message "Tests done!")
+
+(expectations
+  (expect "this is a card" (orgtrello/--card-p `((idList . "this is a card"))))
+  (expect nil (orgtrello/--card-p `((anything-else . "this is not a card")))))
+
+(expectations
+  (expect "this is a checklist" (orgtrello/--checklist-p `((idCard . "this is a checklist"))))
+  (expect nil (orgtrello/--checklist-p `((anything-else . "this is not a checklist")))))
+
+(expectations
+  (expect "this is an item" (orgtrello/--item-p `((state . "this is an item"))))
+  (expect nil (orgtrello/--checklist-p `((anything-else . "this is not a item")))))
+
+(expectations
+ (expect "- [X] call people [4/4]                                        :PROPERTIES: {\"orgtrello-id\":\"456\"}" (orgtrello-cbx/--justify-property-current-line "- [X] call people [4/4] :PROPERTIES: {\"orgtrello-id\":\"456\"}" 100))
+
+ (expect "- [X] call people [4/4]                                         :PROPERTIES: {\"orgtrello-id\":\"456\"}" (orgtrello-cbx/--justify-property-current-line "- [X] call people [4/4]                                         :PROPERTIES: {\"orgtrello-id\":\"456\"}" 100)))
 
 (provide 'org-trello-tests)
 ;;; org-trello-tests ends here