+++ scripts/release.sh
@@ -1,4 +1,4 @@
-#!/bin/bash
+#!/bin/bash -xe
 
 VERSION=$1
 
@@ -17,5 +17,6 @@ git push --tag
 make package
 
 # FIXME
+
 # automate the release to marmalade (rest api)
 # automate the release page from github (rest api too)